Fix build and backspace in vga

+++ src/drivers/pic.rs
@@ -43,7 +43,7 @@ pub unsafe fn remap() {
 
 pub unsafe fn enable(irq: u8) {
   let port: u16 = if (irq & 0b1000) == 0 { 0x21 } else { 0xa1 };
-  let mask: u8 = !(1u8 << (irq & 0b111));
+  let mask: u8 = !(1u8 << (irq & 0b111) as uint);
 
   out(port, inb(port) & mask);
 }
